fix: resolve all dev tool findings — mypy, bandit, test assertions
- Fix mypy errors: type casts for SQLAlchemy Column types
- Fix bandit high: add usedforsecurity=False to md5 cache key hash
- Fix 12 test assertions: wrap Decimal price comparisons with float()
- Update test mocks for claude_client refactor (materials_ai_search, sources)

+++ app/models/price_snapshot.py
@@ -4,7 +4,7 @@ Called by: price_snapshot_service.record_price_snapshot()
 Depends on: MaterialCard (FK)
 """
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Index, Integer, String
+from sqlalchemy import Column, DateTime, ForeignKey, Index, Integer, Numeric, String
 from sqlalchemy.sql import func
 
 from .base import Base
@@ -17,7 +17,7 @@ class MaterialPriceSnapshot(Base):
     id = Column(Integer, primary_key=True)
     material_card_id = Column(Integer, ForeignKey("material_cards.id", ondelete="CASCADE"), index=True, nullable=False)
     vendor_name = Column(String(200), nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Numeric(12, 4), nullable=False)
     currency = Column(String(3), default="USD")
     quantity = Column(Integer, nullable=True)
     source = Column(String(50), nullable=False)
